IMPALA-9162: Do not apply inferred predicate to outer joins
When the planner migrates predicates to inline views, it also creates
equivalent predicates based on the value transfer graph which is built
by transitive relationships among join conditions. These newly inferred
predicates are placed typically as 'other predicates' of an inner or
outer join.

However, for outer joins, this has the effect of adding extra predicates
in the WHERE clause which is incorrect since it may filter NULL values.
Since the original query did not have null filtering conditions in
the WHERE clause, we should not add new ones. This fix does 2 things:
  - At the time of creating the values transfer graph in Analyzer,
    don't add a directed edge between two slots if the target slot
    references the output tuple of an outer join.
  - During the migration of conjuncts to inline views, analyze the
    predicate of type A <op> B and if it is an inferred predicate AND
    either the left or right slots reference the output tuple of an outer
    join, the inferred predicate is ignored. This serves as a safety
    check in case any unqualified predicate 'fell through' until this
    stage of planning.

Note that simple queries with combination of inner and outer joins may
not reproduce the problem.  Due to the nature of predicate inferencing,
some combination of subqueries, inner joins, outer joins is needed.  For
the query pattern, please see the example in the JIRA.

Tests:
  - Added plan tests with left and right outer joins to inline-view.test
  - Manually ran few queries on impala shell to verify result
correctness: by checking that NULL values are being produced for outer
joins.
  - Ran regression tests on jenkins

Change-Id: Ie9521bd768c4b333069c34d5c1e11b10ea535827